+++ src/components/molecules/widgets/MostSevereAccidentsMapWidget.tsx
@@ -1,7 +1,6 @@
-import React, { FC } from 'react';
+import { FC } from 'react';
 import LocationMap from '../LocationMap';
 import { IWidgetMostSevereAccidentsData } from '../../../models/WidgetData';
-import { INITIAL_CENTER } from '../../../utils/utils';
 
 interface IProps {
   data: IWidgetMostSevereAccidentsData;
@@ -9,6 +8,6 @@ interface IProps {
 }
 const MostSevereAccidentsMapWidget: FC<IProps> = ({ data, sizeOptions }) => {
   const { items } = data;
-  return <LocationMap items={items} center={INITIAL_CENTER} sizeOptions={sizeOptions} />;
+  return <LocationMap items={items} />;
 };
 export default MostSevereAccidentsMapWidget;
