+++ src/utils/utils.ts
@@ -46,3 +46,4 @@ export const DEMO_ID = 999;
 export const showOnlyOperCards = process.env.REACT_APP_SHOW_ONLY_OPER_CARDS === 'true';
 
 export const INITIAL_CENTER = { lat: 32.0853, lng: 34.7818 };
+export const INITIAL_ZOOM = parseInt(process.env.REACT_APP_DEFAULT_MAP_ZOOM!);
